test(artifacts): assert (un)expected page-data/html files
