+++ integration-tests/artifacts/src/templates/dummy.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function DummyTemplate({ pageContext }) {
+  return <p>Dummy page for {pageContext.dummyId}</p>
+}
